Bugfixes in /list and /join
Perviously, using /list with no arguments when "the battle has yet to
begin!" would list the correct character's name, but ping the owner of
the last character in the order of initiative.

In addition, if a character was not at max health when they /join'd the
battle, their health would not be set to maximum.

Both of these issues are fixed now.